+++ aider/args.py
@@ -66,6 +66,14 @@ def get_parser(default_config_files, git_root):
         const=sonnet_model,
         help=f"Use {sonnet_model} model for the main chat",
     )
+    haiku_model = "claude-3-5-haiku-20241022"
+    group.add_argument(
+        "--haiku",
+        action="store_const",
+        dest="model",
+        const=haiku_model,
+        help=f"Use {haiku_model} model for the main chat",
+    )
     gpt_4_model = "gpt-4-0613"
     group.add_argument(
         "--4",
@@ -550,6 +558,25 @@ def get_parser(default_config_files, git_root):
     )
 
     ##########
+    group = parser.add_argument_group("Analytics")
+    group.add_argument(
+        "--analytics",
+        action=argparse.BooleanOptionalAction,
+        default=False,
+        help="Enable/disable analytics for one session (default: False)",
+    )
+    group.add_argument(
+        "--analytics-log",
+        metavar="ANALYTICS_LOG_FILE",
+        help="Specify a file to log analytics events",
+    )
+    group.add_argument(
+        "--analytics-disable",
+        action="store_true",
+        help="Permanently disable analytics",
+        default=False,
+    )
+
     group = parser.add_argument_group("Other Settings")
     group.add_argument(
         "--file",
@@ -612,6 +639,12 @@ def get_parser(default_config_files, git_root):
         help="Apply the changes from the given file instead of running the chat (debug)",
     )
     group.add_argument(
+        "--apply-clipboard-edits",
+        action="store_true",
+        help="Apply clipboard contents as edits using the main model's editor format",
+        default=False,
+    )
+    group.add_argument(
         "--yes-always",
         action="store_true",
         help="Always say yes to every confirmation",
@@ -661,6 +694,11 @@ def get_parser(default_config_files, git_root):
         ),
     )
     group.add_argument(
+        "--load",
+        metavar="LOAD_FILE",
+        help="Load and execute /commands from a file on launch",
+    )
+    group.add_argument(
         "--encoding",
         default="utf-8",
         help="Specify the encoding for input and output (default: utf-8)",
@@ -678,8 +716,8 @@ def get_parser(default_config_files, git_root):
     group.add_argument(
         "--gui",
         "--browser",
-        action="store_true",
-        help="Run aider in your browser",
+        action=argparse.BooleanOptionalAction,
+        help="Run aider in your browser (default: False)",
         default=False,
     )
     group.add_argument(