+++ aider/models.py
@@ -13,7 +13,6 @@ import json5
 import yaml
 from PIL import Image
 
-from aider import urls
 from aider.dump import dump  # noqa: F401
 from aider.llm import litellm
 
@@ -53,6 +52,7 @@ ANTHROPIC_MODELS = """
 claude-2
 claude-2.1
 claude-3-haiku-20240307
+claude-3-5-haiku-20241022
 claude-3-opus-20240229
 claude-3-sonnet-20240229
 claude-3-5-sonnet-20240620
@@ -234,24 +234,24 @@ MODEL_SETTINGS = [
     ModelSettings(
         "claude-3-opus-20240229",
         "diff",
-        weak_model_name="claude-3-haiku-20240307",
+        weak_model_name="claude-3-5-haiku-20241022",
         use_repo_map=True,
     ),
     ModelSettings(
         "openrouter/anthropic/claude-3-opus",
         "diff",
-        weak_model_name="openrouter/anthropic/claude-3-haiku",
+        weak_model_name="openrouter/anthropic/claude-3-5-haiku",
         use_repo_map=True,
     ),
     ModelSettings(
         "claude-3-sonnet-20240229",
         "whole",
-        weak_model_name="claude-3-haiku-20240307",
+        weak_model_name="claude-3-5-haiku-20241022",
     ),
     ModelSettings(
         "claude-3-5-sonnet-20240620",
         "diff",
-        weak_model_name="claude-3-haiku-20240307",
+        weak_model_name="claude-3-5-haiku-20241022",
         editor_model_name="claude-3-5-sonnet-20240620",
         editor_edit_format="editor-diff",
         use_repo_map=True,
@@ -268,7 +268,7 @@ MODEL_SETTINGS = [
     ModelSettings(
         "anthropic/claude-3-5-sonnet-20240620",
         "diff",
-        weak_model_name="anthropic/claude-3-haiku-20240307",
+        weak_model_name="anthropic/claude-3-5-haiku-20241022",
         editor_model_name="anthropic/claude-3-5-sonnet-20240620",
         editor_edit_format="editor-diff",
         use_repo_map=True,
@@ -285,7 +285,41 @@ MODEL_SETTINGS = [
     ModelSettings(
         "anthropic/claude-3-5-sonnet-20241022",
         "diff",
-        weak_model_name="anthropic/claude-3-haiku-20240307",
+        weak_model_name="anthropic/claude-3-5-haiku-20241022",
+        editor_model_name="anthropic/claude-3-5-sonnet-20241022",
+        editor_edit_format="editor-diff",
+        use_repo_map=True,
+        examples_as_sys_msg=True,
+        extra_params={
+            "extra_headers": {
+                "anthropic-beta": ANTHROPIC_BETA_HEADER,
+            },
+            "max_tokens": 8192,
+        },
+        cache_control=True,
+        reminder="user",
+    ),
+    ModelSettings(
+        "bedrock/anthropic.claude-3-5-sonnet-20241022-v2:0",
+        "diff",
+        weak_model_name="bedrock/anthropic.claude-3-5-haiku-20241022-v1:0",
+        editor_model_name="bedrock/anthropic.claude-3-5-sonnet-20241022-v2:0",
+        editor_edit_format="editor-diff",
+        use_repo_map=True,
+        examples_as_sys_msg=True,
+        extra_params={
+            "extra_headers": {
+                "anthropic-beta": ANTHROPIC_BETA_HEADER,
+            },
+            "max_tokens": 8192,
+        },
+        cache_control=True,
+        reminder="user",
+    ),
+    ModelSettings(
+        "anthropic/claude-3-5-sonnet-latest",
+        "diff",
+        weak_model_name="anthropic/claude-3-5-haiku-20241022",
         editor_model_name="anthropic/claude-3-5-sonnet-20241022",
         editor_edit_format="editor-diff",
         use_repo_map=True,
@@ -302,7 +336,7 @@ MODEL_SETTINGS = [
     ModelSettings(
         "claude-3-5-sonnet-20241022",
         "diff",
-        weak_model_name="claude-3-haiku-20240307",
+        weak_model_name="claude-3-5-haiku-20241022",
         editor_model_name="claude-3-5-sonnet-20241022",
         editor_edit_format="editor-diff",
         use_repo_map=True,
@@ -329,6 +363,52 @@ MODEL_SETTINGS = [
         cache_control=True,
     ),
     ModelSettings(
+        "anthropic/claude-3-5-haiku-20241022",
+        "diff",
+        weak_model_name="anthropic/claude-3-5-haiku-20241022",
+        use_repo_map=True,
+        extra_params={
+            "extra_headers": {
+                "anthropic-beta": ANTHROPIC_BETA_HEADER,
+            },
+        },
+        cache_control=True,
+    ),
+    ModelSettings(
+        "bedrock/anthropic.claude-3-5-haiku-20241022-v1:0",
+        "diff",
+        weak_model_name="bedrock/anthropic.claude-3-5-haiku-20241022-v1:0",
+        use_repo_map=True,
+        extra_params={
+            "extra_headers": {
+                "anthropic-beta": ANTHROPIC_BETA_HEADER,
+            },
+        },
+        cache_control=True,
+    ),
+    ModelSettings(
+        "claude-3-5-haiku-20241022",
+        "diff",
+        weak_model_name="claude-3-5-haiku-20241022",
+        use_repo_map=True,
+        examples_as_sys_msg=True,
+        extra_params={
+            "extra_headers": {
+                "anthropic-beta": ANTHROPIC_BETA_HEADER,
+            },
+        },
+        cache_control=True,
+    ),
+    ModelSettings(
+        "vertex_ai/claude-3-5-haiku@20241022",
+        "diff",
+        weak_model_name="vertex_ai/claude-3-5-haiku@20241022",
+        use_repo_map=True,
+        extra_params={
+            "max_tokens": 4096,
+        },
+    ),
+    ModelSettings(
         "claude-3-haiku-20240307",
         "whole",
         weak_model_name="claude-3-haiku-20240307",
@@ -343,7 +423,7 @@ MODEL_SETTINGS = [
     ModelSettings(
         "openrouter/anthropic/claude-3.5-sonnet",
         "diff",
-        weak_model_name="openrouter/anthropic/claude-3-haiku",
+        weak_model_name="openrouter/anthropic/claude-3-5-haiku",
         editor_model_name="openrouter/anthropic/claude-3.5-sonnet",
         editor_edit_format="editor-diff",
         use_repo_map=True,
@@ -357,7 +437,7 @@ MODEL_SETTINGS = [
     ModelSettings(
         "openrouter/anthropic/claude-3.5-sonnet:beta",
         "diff",
-        weak_model_name="openrouter/anthropic/claude-3-haiku:beta",
+        weak_model_name="openrouter/anthropic/claude-3-5-haiku:beta",
         editor_model_name="openrouter/anthropic/claude-3.5-sonnet:beta",
         editor_edit_format="editor-diff",
         use_repo_map=True,
@@ -373,7 +453,7 @@ MODEL_SETTINGS = [
     ModelSettings(
         "vertex_ai/claude-3-5-sonnet@20240620",
         "diff",
-        weak_model_name="vertex_ai/claude-3-haiku@20240307",
+        weak_model_name="vertex_ai/claude-3-5-haiku@20241022",
         editor_model_name="vertex_ai/claude-3-5-sonnet@20240620",
         editor_edit_format="editor-diff",
         use_repo_map=True,
@@ -386,7 +466,7 @@ MODEL_SETTINGS = [
     ModelSettings(
         "vertex_ai/claude-3-5-sonnet-v2@20241022",
         "diff",
-        weak_model_name="vertex_ai/claude-3-haiku@20240307",
+        weak_model_name="vertex_ai/claude-3-5-haiku@20241022",
         editor_model_name="vertex_ai/claude-3-5-sonnet-v2@20241022",
         editor_edit_format="editor-diff",
         use_repo_map=True,
@@ -399,13 +479,13 @@ MODEL_SETTINGS = [
     ModelSettings(
         "vertex_ai/claude-3-opus@20240229",
         "diff",
-        weak_model_name="vertex_ai/claude-3-haiku@20240307",
+        weak_model_name="vertex_ai/claude-3-5-haiku@20241022",
         use_repo_map=True,
     ),
     ModelSettings(
         "vertex_ai/claude-3-sonnet@20240229",
         "whole",
-        weak_model_name="vertex_ai/claude-3-haiku@20240307",
+        weak_model_name="vertex_ai/claude-3-5-haiku@20241022",
     ),
     # Cohere
     ModelSettings(
@@ -467,6 +547,11 @@ MODEL_SETTINGS = [
     ),
     ModelSettings(
         "gemini/gemini-1.5-pro-exp-0827",
+        "diff-fenced",
+        use_repo_map=True,
+    ),
+    ModelSettings(
+        "vertex_ai/gemini-pro-experimental",
         "diff-fenced",
         use_repo_map=True,
     ),
@@ -630,79 +715,91 @@ MODEL_SETTINGS = [
         use_temperature=False,
         streaming=False,
     ),
+    ModelSettings(
+        "openrouter/qwen/qwen-2.5-coder-32b-instruct",
+        "diff",
+        weak_model_name="openrouter/qwen/qwen-2.5-coder-32b-instruct",
+        editor_model_name="openrouter/qwen/qwen-2.5-coder-32b-instruct",
+        editor_edit_format="editor-diff",
+        use_repo_map=True,
+    ),
 ]
 
 
-model_info_url = (
-    "https://raw.githubusercontent.com/BerriAI/litellm/main/model_prices_and_context_window.json"
-)
+class ModelInfoManager:
+    MODEL_INFO_URL = (
+        "https://raw.githubusercontent.com/BerriAI/litellm/main/"
+        "model_prices_and_context_window.json"
+    )
+    CACHE_TTL = 60 * 60 * 24  # 24 hours
 
+    def __init__(self):
+        self.cache_dir = Path.home() / ".aider" / "caches"
+        self.cache_file = self.cache_dir / "model_prices_and_context_window.json"
+        self.content = None
+        self._load_cache()
 
-def get_model_flexible(model, content):
-    info = content.get(model, dict())
-    if info:
-        return info
-
-    pieces = model.split("/")
-    if len(pieces) == 2:
-        info = content.get(pieces[1])
-        if info and info.get("litellm_provider") == pieces[0]:
-            return info
-
-    return dict()
-
-
-def get_model_info(model):
-    if not litellm._lazy_module:
-        cache_dir = Path.home() / ".aider" / "caches"
-        cache_file = cache_dir / "model_prices_and_context_window.json"
-
+    def _load_cache(self):
         try:
-            cache_dir.mkdir(parents=True, exist_ok=True)
-            use_cache = True
+            self.cache_dir.mkdir(parents=True, exist_ok=True)
+            if self.cache_file.exists():
+                cache_age = time.time() - self.cache_file.stat().st_mtime
+                if cache_age < self.CACHE_TTL:
+                    self.content = json.loads(self.cache_file.read_text())
         except OSError:
-            # If we can't create the cache directory, we'll skip using the cache
-            use_cache = False
+            pass
 
-        if use_cache:
-            current_time = time.time()
-            cache_age = (
-                current_time - cache_file.stat().st_mtime if cache_file.exists() else float("inf")
-            )
-
-            if cache_age < 60 * 60 * 24:
-                try:
-                    content = json.loads(cache_file.read_text())
-                    res = get_model_flexible(model, content)
-                    if res:
-                        return res
-                except Exception as ex:
-                    print(str(ex))
-
-        import requests
-
+    def _update_cache(self):
         try:
-            response = requests.get(model_info_url, timeout=5)
+            import requests
+
+            response = requests.get(self.MODEL_INFO_URL, timeout=5)
             if response.status_code == 200:
-                content = response.json()
-                if use_cache:
-                    try:
-                        cache_file.write_text(json.dumps(content, indent=4))
-                    except OSError:
-                        # If we can't write to the cache file, we'll just skip caching
-                        pass
-                res = get_model_flexible(model, content)
-                if res:
-                    return res
+                self.content = response.json()
+                try:
+                    self.cache_file.write_text(json.dumps(self.content, indent=4))
+                except OSError:
+                    pass
         except Exception as ex:
             print(str(ex))
 
-    # If all else fails, do it the slow way...
-    try:
-        info = litellm.get_model_info(model)
-        return info
-    except Exception:
+    def get_model_from_cached_json_db(self, model):
+        if not self.content:
+            self._update_cache()
+
+        if not self.content:
+            return dict()
+
+        info = self.content.get(model, dict())
+        if info:
+            return info
+
+        pieces = model.split("/")
+        if len(pieces) == 2:
+            info = self.content.get(pieces[1])
+            if info and info.get("litellm_provider") == pieces[0]:
+                return info
+
         return dict()
+
+    def get_model_info(self, model):
+        cached_info = self.get_model_from_cached_json_db(model)
+
+        litellm_info = None
+        if litellm._lazy_module or not cached_info:
+            try:
+                litellm_info = litellm.get_model_info(model)
+            except Exception as ex:
+                if "model_prices_and_context_window.json" not in str(ex):
+                    print(str(ex))
+
+        if litellm_info:
+            return litellm_info
+
+        return cached_info
+
+
+model_info_manager = ModelInfoManager()
 
 
 class Model(ModelSettings):
@@ -737,7 +834,7 @@ class Model(ModelSettings):
             self.get_editor_model(editor_model, editor_edit_format)
 
     def get_model_info(self, model):
-        return get_model_info(model)
+        return model_info_manager.get_model_info(model)
 
     def configure_model_settings(self, model):
         for ms in MODEL_SETTINGS:
@@ -777,6 +874,22 @@ class Model(ModelSettings):
             self.use_repo_map = True
             self.examples_as_sys_msg = True
             self.reminder = "user"
+
+        if model.startswith("o1-") or "/o1-" in model:
+            self.use_system_prompt = False
+            self.use_temperature = False
+            self.streaming = False
+
+        if (
+            "qwen" in model
+            and "coder" in model
+            and ("2.5" in model or "2-5" in model)
+            and "32b" in model
+        ):
+            "openrouter/qwen/qwen-2.5-coder-32b-instruct",
+            self.edit_format = "diff"
+            self.editor_edit_format = "editor-diff"
+            self.use_repo_map = True
 
         # use the defaults
         if self.edit_format == "diff":
@@ -964,8 +1077,14 @@ def register_litellm_models(model_fnames):
             continue
 
         try:
-            with open(model_fname, "r") as model_def_file:
-                model_def = json5.load(model_def_file)
+            data = Path(model_fname).read_text()
+            if not data.strip():
+                continue
+            model_def = json5.loads(data)
+            if not model_def:
+                continue
+
+            # only load litellm if we have actual data
             litellm._load_litellm()
             litellm.register_model(model_def)
         except Exception as e:
@@ -1037,9 +1156,6 @@ def sanity_check_model(io, model):
             for match in possible_matches:
                 io.tool_output(f"- {match}")
 
-    if show:
-        io.tool_output(f"For more info, see: {urls.model_warnings}")
-
     return show
 
 
@@ -1051,7 +1167,10 @@ def fuzzy_match_models(name):
         model = model.lower()
         if attrs.get("mode") != "chat":
             continue
-        provider = (attrs["litellm_provider"] + "/").lower()
+        provider = attrs.get("litellm_provider", "").lower()
+        if not provider:
+            continue
+        provider += "/"
 
         if model.startswith(provider):
             fq_model = model